MDL_CLASS_COUNT is long
git-svn-id: file:///home/roberto/Git/svn@11656 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: a7ea0ae0d5fe8a8291da04bcad8ab13ee32a6286

+++ trunk/xmipp/libraries/parallel/mpi_classify_FTTRI.cpp
@@ -1063,7 +1063,7 @@ void ProgClassifyFTTRI::writeResults(bool FTTRI)
     {
         MDclass.clear();
         const std::vector<size_t> &class_i=bestEpsilonClasses[i].memberIdx;
-        int nmax=class_i.size();
+        size_t nmax=class_i.size();
 
         // Write this class in summary
         size_t id=MDsummary.addObject();
